pyproject.toml
[project]
name = "temporal-agentic-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.13",
    "click>=8.3.1",
    "fastapi>=0.135.2",
    "psycopg2-binary>=2.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "temporalio==1.10.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "."]
markers = [
    "integration: marks tests as integration tests (require running Temporal server)",
    "llm: marks tests that call a real LLM (run with --llm flag)",
]
